+++ code/edgemedia/src/main/java/com/adobe/marketing/mobile/edge/media/internal/MediaXDMEventHelper.kt
@@ -193,8 +193,8 @@ internal class MediaXDMEventHelper {
             return XDMChapterDetails(
                 chapterInfo.name,
                 chapterInfo.position,
-                chapterInfo.length.toLong(),
-                chapterInfo.startTime.toLong()
+                chapterInfo.length,
+                chapterInfo.startTime
             )
         }
