+++ code/edgemedia/src/main/java/com/adobe/marketing/mobile/edge/media/internal/xdm/XDMChapterDetails.kt
@@ -13,9 +13,9 @@ package com.adobe.marketing.mobile.edge.media.internal.xdm
 
 internal data class XDMChapterDetails(
     var friendlyName: String? = null,
-    var index: Long? = null,
-    var length: Long? = null,
-    var offset: Long? = null
+    var index: Int? = null,
+    var length: Int? = null,
+    var offset: Int? = null
 ) : XDMProperty {
 
     override fun serializeToXDM(): Map<String, Any> {
